+++ fabric/src/main/java/com/lx862/jcm/mod/render/text/VanillaTextRenderer.java
@@ -10,18 +10,18 @@ import org.mtr.mapping.mapper.GraphicsHolder;
  * Wrapper around vanilla text rendering
  */
 public class VanillaTextRenderer implements RenderHelper {
-    public static void draw(GraphicsHolder graphicsHolder, TextInfo text, int x, int y) {
+    public static void draw(GraphicsHolder graphicsHolder, TextInfo text, double x, double y) {
         drawInternal(graphicsHolder, text, x, y);
     }
 
-    private static void drawInternal(GraphicsHolder graphicsHolder, TextInfo text, int x, int y) {
+    private static void drawInternal(GraphicsHolder graphicsHolder, TextInfo text, double x, double y) {
         int textWidth = GraphicsHolder.getTextWidth(text.toMutableText());
         double finalX = text.getTextAlignment().getX(x, textWidth);
         MutableText finalText = text.toMutableText();
-        graphicsHolder.drawText(finalText, (int)finalX, y, text.getTextColor(), false, MAX_RENDER_LIGHT);
+        graphicsHolder.drawText(finalText, (int)finalX, (int)y, text.getTextColor(), false, MAX_RENDER_LIGHT);
     }
 
-    public static void drawScrollingText(GraphicsHolder graphicsHolder, String text, int maxW, int x, int y, int textColor) {
+    public static void drawScrollingText(GraphicsHolder graphicsHolder, String text, int maxW, double x, double y, int textColor) {
         String str = text;
         int scrollSpeed = str.length() * 6;
         int fullTick = (JCMStats.getGameTick() % (int)(scrollSpeed * 1.5));
@@ -64,7 +64,7 @@ public class VanillaTextRenderer implements RenderHelper {
             graphicsHolder.translate(-((prg - start) * nextCharWidth), 0, 0);
         }
 
-        graphicsHolder.drawText(newText, x, y, textColor, false, MAX_RENDER_LIGHT);
+        graphicsHolder.drawText(newText, (int)x, (int)y, textColor, false, MAX_RENDER_LIGHT);
         graphicsHolder.pop();
     }
